Abaco

Abaco computes formulas output from a stream of input variables.

Abaco aims to target the case where a number of IoT sensor, network devices or other types of objects that send periodic variable values must be combined into output values defined by math expressions.

In a real world scenario values are coming asynchronously, delayed and out of orders. Abaco may manage values referring to different times.

Basic Concepts

formula

A named math expression:

    myformula = x + y

As soon as all inputs variables, x and y, are collected the formula myformula is evaluated and the result available to the onresult callback.

A formula is computable if all the independents variables belong to the same time window.

sn

The identifier of the element sending the input variables.

It is up to you to choose the acronym for sn that best fits the context: sensor name, scope name or short name if it were a unique string extracted for a long distinguish name dn string.

time window

A time window includes the all the timestamps in the interval rop:

rop = { t ∈ N | START_INTERVAL <= t < END_TIME }

Timestamps are integer values with second granularity.

In a periodic data collection system the time window rop identifies a Report Output Period.

For exammple suppose that a data collection system uses a 15 minutes rop interval: in this case an hour is divided into 4 rops and from ten to eleven of some (omitted) day you have:

rop1 = { t ∈ [10:00:00, 10:15:00) }

rop2 = { t ∈ [10:15:00, 10:30:00) }

rop3 = { t ∈ [10:30:00, 10:45:00) }

rop4 = { t ∈ [10:45:00, 11:00:00) }

Conventionally the rop is identified by its START_TIME: in the above example rop1 is the ten o'clock rop.

The rop interval width is an abaco setting, user-defined at startup.

ages

The number of ages defines how many consecutive rops are managed.

For example, for a rop width of 15 minutes, set ages to 4 if your network devices may send data with a maximum delay of an hour. A received value marked with a timestamp distant 4 or more rops from the latest rop is discarded.

The number of ages is an abaco setting, user-defined at startup.

This is the minimal background theory, the below example should help to clarify the Abaco mechanics.

Getting Started

Installation:

julia> using Pkg; Pkg.add("Abaco")    

Usage:

using Abaco

# Initialize abaco context with a rop time_window of 60 seconds and handle
# input values with timestamp ts up to 4 contiguous rops.
abaco = abaco_init(width=60, ages=4) do ts, sn, fname, value, inputs
    @info "[$ts][$sn] function $fname=$value"
end

# Add desired outputs in terms of inputs variables x, y, z, v, w
outputs = ["xysum = x + y", "rsigma = x * exp(y-1)", "wsum = (x*w + z*v)"]

for formula in outputs
    add_formula!(abaco, formula)
end

# Start receiving some inputs values
# normally ts is the UTC timestamp from epoch in seconds.
# but for semplicity assume time start from zero.

# the device AG101 sends the x value at timestamp 0.
ts = 0
device = "AG101"
add_value!(abaco, ts, device, "x", 10)

# Time flows and about 1 minute later ...

# the device CE987 sends the y value at timestamp 65.
ts = 65
device = "CE987"
add_value!(abaco, ts, device, "y", 10)

# Time flows and more than 1 minute later ...

# Finally the device AG101 sends the y value calculated at timestamp 0.
# At this instant the formulas that depends on x and y are computable
# for the element AG101 at timestamp 0.
ts = 0
device = "AG101"
add_value!(abaco, ts, device, "y", 20)
[ Info: [0][AG101] function xysum=30
[ Info: [0][AG101] function rsigma=1.7848230096318724e9

# Now arrives the variable x from CE987 that make some formulas computables.
# Note that x timestamp is 65, y timestamp is 101
# and the formulas timestamp is 60: the START_TIME of the rop. 
ts = 101
device = "CE987"
add_value!(abaco, ts, device, "x", 10)
[ Info: [60][CE987] function xysum=20
[ Info: [60][CE987] function rsigma=81030.83927575384

In the formula callback the first argument may be a user defined object, like a Socket a Channel or whatsoever communication endpoint.

For example:

sock = connect(3001)

abaco = abaco_init(handle=sock, width=900) do sock, ts, sn, name, value, inputs
    @info "age [$ts]: [$sn] $name = $value"
    msg = JSON3.write(Dict(
                        "sn" => sn,
                        "age" => ts, 
                        "formula" => name,
                        "value" => value))
    write(sock, msg*"\n")
end

API

Abaco.Context

Abaco.EvalError

Abaco.WrongFormula

Abaco.abaco_init

Abaco.add_formula!

Abaco.add_value!

Abaco.add_values!

Abaco.add_values!

Abaco.nowts

Abaco.rop

Abaco.ropcurrent

Abaco.Context — Type

Maintains the state of the abaco.

Before adding formulas and values an abaco Context must be initialized by abaco_init.

Abaco.EvalError — Type

Formula evaluation failure.

[add_value!] throws EvalError when a runtime formula evaluation fails, for example for a wrong numbers of method args:

add_formula!(abaco, "div(x,y,z")
add_value!(abaco, ts, sn, Dict("x"=>10, "y"=>1, "z"=1))

Abaco.WrongFormula — Type

Wrong formula definition.

add_formula! throws WrongFormula when a formula is malformed, for example:

add_formula!(abaco, "myformula = x + ")

Abaco.abaco_init — Method

abaco_init(onresult; handle=nothing, width::Int=900, ages::Int=4)::Context

Initialize the abaco context:

onresult: function callback that gets called each time a formula value is computed.

handle: user defined object. If handle is defined it is the first argument of onresult, default to nothing.

width: the rop interval in seconds, default to 900 seconds (15 minutes).

ages: the number of active rops managed by the abaco, default to 4.

Example 1: defining onresult callback that uses of an handle object.

# the handle object is a socket
sock = connect(3001)

function onresult(handle, ts, sn, formula_name, value, inputs)
    # build a pkt message from ts, sn, ...
    pkt = ...
    write(sock, pkt)
end

Example 2: defining onresult callback that doesn't use an handle object.

abaco = abaco_init(onresult, handle=sock)

function onresult(ts, sn, formula_name, value, inputs)
    @info "[$ts][$sn] function $fname=$value"
end

abaco = abaco_init(onresult)

Abaco.add_formula! — Method

add_formula!(abaco::Context, formula_def::String)

Add a formula, with formula_def formatted as "formula_name = expression", where expression is a mathematical expression, like x + y*w.

Abaco.add_value! — Method

add_value!(abaco, ts::Int, sn::String, var::String, val::Real)

Adds the input variable var with value val.

ts: timestamp with seconds resolution

sn: scope name

var: variable name

val: variable value

Abaco.add_values! — Method

add_values!(abaco, ts, sn, values)

Adds the input variables include in the dictionary values.

    # now timestamp 
    ts = nowts()

    # short name of network element
    sn = "trento.castello"

    values = Dict(
        "x" => 23.2,
        "y" => 100
    )
    add_values(abaco, ts, sn, values)

Abaco.add_values! — Method

add_values!(abaco, payload)

Adds the input variables included in the payload dictionary.

The Dict msg must contains the keys ts and sn and a numbers of other keys managed as input variables.

This function modifies the payload dictionary: sn and ts keys are popped out.

    payload = Dict(
        "ts" => nowts(),
        "sn" => "trento.castello",
        "x" => 23.2,
        "y" => 100
    )
    add_values(abaco, ts, sn, payload)

Abaco.nowts — Function

nowts()

the current timestamp in seconds from epoch

Abaco.rop — Function

rop(ts::Int64, width::Int64=900)

Returns the start_time of the rop time window that contains ts.

The ROP is the Report Output Period of a metric and it is the time interval defined as:

ROP = { t ∈ [start_time, start_time+width) }

It is customary to use the start_time to indicate a rop: the rop of ten 'oclock is the time interval [10:00, 10:15).

Abaco.ropcurrent — Function

ropcurrent(interval::Int64=900)::Int64

Returns the epoch start time of the current rop. The current rop is the nearest in the present rop that satisfies the condition: rop.endtime < now.